requirements.txt
telethon>=1.36
anthropic>=0.39
tweepy>=4.14
python-dotenv>=1.0
httpx>=0.27
